# backend/storage/api-docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CMS Project API",
        "description": "    \n    ⚠️ Authorization Guide:\n    1. Use API **Login** to get Access Token.\n    2. Click on **[Authorize]** button.\n    3. Paste the token into the field.\n    4. If you receive **401**, it means you haven't logged in or the token has expired.\n    5. If you receive **403**, it means you don't have permission to access.\n\n    *Note: \n    - If you haven't logged in but the system still run the API (due to browser saving cookie), try again with anonymous.\n    - The System verify the token through cookie.\n    ",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "description": "Login user",
                "operationId": "cba39ed8f23a8cde6b7228dd2e559187",
                "requestBody": {
                    "$ref": "#/components/requestBodies/AuthsLogin"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/AuthsLogin"
                    },
                    "401": {
                        "$ref": "#/components/responses/Exception401"
                    }
                }
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout user",
                "description": "Logout user",
                "operationId": "4cdd245e58e1b22a0119bd159fb8e5c1",
                "responses": {
                    "203": {
                        "$ref": "#/components/responses/AuthsLogout"
                    },
                    "401": {
                        "$ref": "#/components/responses/Exception401"
                    }
                }
            }
        },
        "/api/v1/admin/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get current user",
                "description": "Get current user",
                "operationId": "faf8c23a3e9714860d50e779c0b8284d",
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/AuthsMe"
                    },
                    "401": {
                        "$ref": "#/components/responses/Exception401"
                    }
                }
            }
        },
        "/api/v1/password/forgot": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Send reset password link to user",
                "description": "Send reset password link to user",
                "operationId": "4fdd79859330e3cc5080fd57b91e7dfb",
                "requestBody": {
                    "$ref": "#/components/requestBodies/AuthsForgotPassword"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/AuthsForgotPassword"
                    },
                    "400": {
                        "$ref": "#/components/responses/Exception400"
                    },
                    "401": {
                        "$ref": "#/components/responses/Exception401"
                    },
                    "404": {
                        "$ref": "#/components/responses/Exception404"
                    },
                    "500": {
                        "$ref": "#/components/responses/Exception500"
                    }
                }
            }
        },
        "/api/v1/password/reset": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Reset password",
                "description": "Reset password",
                "operationId": "911590cd94386d62e8177dc93a9cf336",
                "requestBody": {
                    "$ref": "#/components/requestBodies/AuthsResetPassword"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/AuthsResetPassword"
                    },
                    "400": {
                        "$ref": "#/components/responses/Exception400"
                    },
                    "401": {
                        "$ref": "#/components/responses/Exception401"
                    },
                    "404": {
                        "$ref": "#/components/responses/Exception404"
                    },
                    "500": {
                        "$ref": "#/components/responses/Exception500"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Exception": {
                "required": [
                    "errorCode",
                    "errorMessage",
                    "data"
                ],
                "properties": {
                    "code": {
                        "type": "string",
                        "example": "ERROR_CODE"
                    },
                    "message": {
                        "type": "string",
                        "example": "Error message."
                    },
                    "errors": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        },
                        "example": []
                    }
                },
                "type": "object"
            },
...
            "Roles": {
                "required": [
                    "role_name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "uniqueItems": true,
                        "example": "admin"
                    },
                    "description": {
                        "type": "string",
                        "example": "web"
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2023-01-01 00:00:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2023-01-01 00:00:00"
                    },
                    "permissions": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        },
                        "example": []
                    }
                },
                "type": "object"
            },
            "Users": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "5d1e8b2a-0a2b-4c0e-9a1b-f6b9a56a9a11"
                    },
                    "user_name": {
                        "type": "string",
                        "example": "jdoe"
                    },
                    "display_name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T12:00:00Z",
                        "nullable": true
                    },
                    "password": {
                        "type": "string",
                        "writeOnly": true,
                        "example": "password123"
                    },
                    "address": {
                        "type": "string",
                        "example": "123 Main St"
                    },
                    "phone": {
                        "type": "string",
                        "example": "1234567890"
                    },
                    "bio": {
                        "type": "string",
                        "example": "A passionate developer",
                        "nullable": true
                    },
                    "avatar": {
                        "type": "string",
                        "example": "avatar.jpg",
                        "nullable": true
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "role_id": {
                        "type": "string",
                        "example": "1"
                    },
                    "role": {
                        "$ref": "#/components/schemas/Roles"
                    },
                    "permissions": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "AuthsForgotPassword": {
                "description": "OK",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "data": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "We have e-mailed your password reset link!"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "AuthsLogin": {
                "description": "OK",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "data": {
                                    "$ref": "#/components/schemas/Users"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "AuthsLogout": {
                "description": "OK"
            },
            "AuthsMe": {
                "description": "OK",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "data": {
                                    "$ref": "#/components/schemas/Users"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "AuthsResetPassword": {
                "description": "OK",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "data": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Password reset successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "Exception400": {
                "description": "Bad Request",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "data": {
                                    "$ref": "#/components/schemas/Exception"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "Exception401": {
                "description": "Unauthorized",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "data": {
                                    "$ref": "#/components/schemas/Exception"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "Exception404": {
                "description": "Not Found",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "data": {
                                    "$ref": "#/components/schemas/Exception"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "Exception500": {
                "description": "Internal Server Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "data": {
                                    "$ref": "#/components/schemas/Exception"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "requestBodies": {
            "AuthsForgotPassword": {
                "description": "User forgot password credentials, email is required",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "email": {
                                    "type": "string",
                                    "example": "admin@example.com"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "AuthsLogin": {
                "description": "User login credentials, email or username is required",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "email": {
                                    "type": "string",
                                    "example": "admin@example.com"
                                },
                                "username": {
                                    "type": "string",
                                    "example": "admin"
                                },
                                "password": {
                                    "type": "string",
                                    "example": "123456"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "AuthsResetPassword": {
                "description": "User reset password credentials, email is required",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "email": {
                                    "type": "string",
                                    "example": "admin@example.com"
                                },
                                "password": {
                                    "type": "string",
                                    "example": "123456"
                                },
                                "password_confirmation": {
                                    "type": "string",
                                    "example": "123456"
                                },
                                "token": {
                                    "type": "string",
                                    "example": "123456"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}